close the two dtype-rule violations the last commit recorded
Both were found while reviewing the interface pass and left out of it because
they sit outside what that change touched. Neither is large; leaving a written
rule contradicted by shipped code is worse than the break.

`matches_above`' indices are int64, not uint32. They are positions along the
scanned axis of the `scores` matrix the caller passed in — the same quantity
`top_matches` returns, which was already int64. Two functions in one module
disagreeing about the dtype of one thing is exactly what the rule exists to
stop. The ragged array is the larger of the two, so this does cost memory; being
able to move an index between the two calls without thinking is worth more.

`MatchError::AxisTooLong` goes with it. It refused a scanned axis longer than
u32::MAX, which was a constraint of the index width and nothing else, so keeping
an error that says "exceeds the u32 index range" when no u32 index remains would
be actively misleading.

`dag.parent_dist` is exported. It was public, docstring'd, listed in the
capability table and used by the test suite, but named in no `__all__` — so
`fastcore.parent_dist` raised AttributeError and only `fastcore.dag.parent_dist`
worked. Its `-> None` annotation was wrong too; it returns the distances. The
capability table said "— (see parent_dist)" for the Python column, which is now
just the function name.

matches is Rust + Python only, so the R bindings are unaffected. The vendored R
copy is refreshed so the drift gate passes.

+++ py/tests/test_matches.py
@@ -190,7 +190,9 @@ def test_matches_above_vs_reference(dtype, axis, distances, crit):
     assert offsets[0] == 0
     assert offsets[-1] == len(indices) == len(values)
     assert (np.diff(offsets) >= 0).all()
-    assert indices.dtype == np.uint32
+    # Positions in `scores`, so int64 — the same dtype `top_matches` uses for the
+    # same quantity. See docs/python/index.md, "Integer return dtypes".
+    assert indices.dtype == np.int64
     assert offsets.dtype == np.int64
     assert values.dtype == dtype
